move get_xsection_script from klayout to gdsfactory.geometry
